added html elements for score and js score logic
added divs to receive text as a visual indicator of score
changes to winner variable , from a string of text declaring a winner
to a shorter string used to test who won.
will implement a new alert for winner next commit.
added function getPoint and testWinner to test if the score of player
or computer has reached 5 and declare a match winner.